First step to single-activity structure
