+++ .gitignore
@@ -1,5 +1,6 @@
 *.iml
 .gradle
+.kotlin
 /local.properties
 /.idea/caches
 /.idea/libraries
@@ -15,7 +16,5 @@
 local.properties
 # Android Studio
 .idea/
-*.iml
 .gradle/
-local.properties
 *.keystore
